requirements.txt
# Core dependencies
fastapi==0.68.0
uvicorn==0.15.0
websockets==10.0

# Google Cloud APIs
google-cloud-vision==2.4.2
google-cloud-speech==2.9.1

# Agent framework
spade==3.2.2

# Testing
pytest==6.2.5
httpx==0.19.0  # For testing FastAPI

# Additional utilities
python-multipart==0.0.5
pydantic==1.8.2
